fix(runner): logs get lost due to lack of Node ID in MDC (#183)
The `mdcPropertyMap` is initialized lazily. Therefore, doing so within the `CoroutineScope` can result in the MDC map gotten being incorrect somehow. Calling the `mdcPropertyMap` getter before entering the `scope` would have worked as well, however, instead, all non-suspending code was moved out of the `scope.launch` to improve throughput with larger scales.

Previously, the idea was to delegate all log code to the log recorder `dispatcher` to not lose any time on constructing the entry while running the Flow. However, when scaling, this single thread that is already busy with the DB I/O will probably be the bottleneck. There's no point in moving a few nanoseconds worth of computations into that Thread.

Fixes #183